data diri penyuplai created

+++ project-ecoranger/Models/Penyuplai.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,7 +11,7 @@ namespace project_ecoranger.Models
     {
         public string idPenyuplai { get; set; }
         public string nama { get; set; }
-        public int noTelp { get; set; }
+        public BigInteger noTelp { get; set; }
         public string alamat { get; set; }
         public string email { get; set; }
         public string username { get; set; }
